requirements.txt
fastapi
httpx
pydantic
markdown
uvicorn
python-dotenv